+++ src/runner/TestRunner.ts
@@ -40,37 +40,37 @@ export class TestRunner {
     ) {
         // Create an instance of the BrightScript TAP object so we can pass it to the tests for reporting.
         let tap = execute(
-            [path.join(__dirname, "..", "resources", "tap.brs")],
+            [path.join(__dirname, "..", "..", "resources", "tap.brs")],
             [new BrsTypes.Int32(testFiles.length)]
         );
-
         let executeArgs = this.generateExecuteArgs(tap, focusedCasesDetected);
-        testFiles.forEach((filename, index) => {
-            this.executeFile(execute, executeArgs, filename);
-            // Update the index so that our TAP reporting is correct.
-            executeArgs.elements.set("index", new BrsTypes.Int32(index + 1));
-        });
+        this._run(execute, executeArgs, testFiles);
     }
 
     /**
-     * Executes and reports a given test file.
+     * Executes and reports a given test file. Subclasses should override this method
+     * with custom implementations as needed.
      * @param execute The scoped execution function
      * @param executeArgs Args to pass to the execution function
      * @param filename The file to execute
      */
-    protected executeFile(
+    protected _run(
         execute: ExecuteWithScope,
         executeArgs: BrsTypes.RoAssociativeArray,
-        filename: string
+        testFiles: string[]
     ) {
-        try {
-            execute([filename], [executeArgs]);
-        } catch (e) {
-            console.error(
-                `Stopping execution. Interpreter encountered an error in ${filename}.`
-            );
-            process.exit(1);
-        }
+        testFiles.forEach((filename, index) => {
+            // Set the index so that our TAP reporting is correct.
+            executeArgs.elements.set("index", new BrsTypes.Int32(index));
+            try {
+                execute([filename], [executeArgs]);
+            } catch (e) {
+                console.error(
+                    `Stopping execution. Interpreter encountered an error in ${filename}:\n\t${e}`
+                );
+                process.exit(1);
+            }
+        });
     }
 
     /**
